+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Change main package name to datadope_alerta
 * Blackouts management plugin
 * GCHAT notifier plugin
+* Telegram notifier plugin
 * Contextualizer API to manage (read, add, delete, update) the rules
   that will be used by the Notifier plugin.
 * Notifier plugin to compare the rules stored in the database with